Projects/team/Software Engineering/drp/src/web/app/requirements.txt
flask
mysql-connector-python
python-dotenv
bokeh
pandas
bcrypt
gunicorn
arrow
statsmodels
flask-cors
redis[hiredis]
pythonping